Creating react apps: Rename and make markdown
Rename `Creating react apps.txt` to
`creating-react-apps.md`, removing spaces in
the name to make it easier to type. Make it
a markdown file, and add markdown formatting.

Later we could collect notes we make to a wiki
in the GitHub repository - but in the repo is
good for now.