pbft: add an 'authentic' flag at servers
Messages are tagged with the 'authentic' flag of the server.
If a server receives a message with 'authentic' set to false,
then the message is not accepted.

This can simulate a server having an invalid key, and the other
servers have some way to tell that the key is invalid.