pbft: use client signature to verify pre-prepares
Since a valid pre-prepare message must have been
signed with a valid client signature, nonfaulty
replicas can detect a Byzantine primary that
attempts to modify the message content sent with
the pre-prepare and overwrites the digest.

Nonfaulty primaries will not accept a pre-prepare
whose message has been modified by a server
with isInByzantineMode set to true. This means
in order for the Byzantine message to be
committed, at least 2f+1 (3 in the current
setup of 5 servers) must have isInByzantineMode
set true.